+++ domain/src/main/java/woowacourse/shopping/repository/CartRepository.kt
@@ -5,5 +5,5 @@ import woowacourse.shopping.model.Product
 interface CartRepository {
     suspend fun addCartProduct(product: Product)
     suspend fun getAllCartProducts(): List<Product>
-    suspend fun deleteCartProduct(id: Int)
+    suspend fun deleteCartProduct(id: Long)
 }
